# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: Luther API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      For accessing the API a valid JWT token must be passed in all the queries in
      the 'Authorization' header.


      A valid JWT token is generated by the API and retourned as answer of a call
      to the route /signin giving a valid user & password.


      The following syntax must be used in the 'Authorization' header :

        xxxxxx.yyyyyyy.zzzzzz
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /auction/getBestDealership:
    x-swagger-router-controller: auction
    get:
      description: Get Best Dealership
      tags:
        - auction
      operationId: getBestDealership
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
...
  /market/getMarketData:
    x-swagger-router-controller: market
    get:
      description: Get Market Data
      tags:
        - market
      operationId: getMarketData
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
  /market/getMakeData:
    x-swagger-router-controller: market
    get:
      description: Get Make Data
      tags:
        - market
      operationId: getMakeData
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
  /market/getModelData:
    x-swagger-router-controller: market
    get:
      description: Get Model Data
      tags:
        - market
      operationId: getModelData
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
  /market/getDealershipData:
    x-swagger-router-controller: market
    get:
      description: Get Dealership Data
      tags:
        - market
      operationId: getDealershipData
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
  /market/getSalesData:
    x-swagger-router-controller: market
    get:
      description: Get Sales Data
      tags:
        - market
      operationId: getSalesData
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
  /market/getLeadsData:
    x-swagger-router-controller: market
    get:
      description: Get Leads Data
      tags:
        - market
      operationId: getLeadsData
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
  /market/getViewsData:
    x-swagger-router-controller: market
    get:
      description: Get Views Data
      tags:
        - market
      operationId: getViewsData
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
  /market/getRevenueData:
    x-swagger-router-controller: market
    get:
      description: Get Revenue Data
      tags:
        - market
      operationId: getRevenueData
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
  /market/getInventoryData:
    x-swagger-router-controller: market
    get:
      description: Get Inventory Data
      tags:
        - market
      operationId: getInventoryData
      security:
        - Bearer: []
      responses:
        '200':
          description: Success
# complex objects have schema definitions
...